On long click of mobile phone or email, it should be copied to clipboard and toast message displayed Enable Call and SMS feature

+++ app/src/main/java/scoproject/com/contactsappgojek/networking/NetworkService.java
@@ -18,5 +18,5 @@ public interface NetworkService {
     Observable<List<People>> getPeopleList();
 
     @GET("/contacts/{id}.json")
-    Observable<People> getPeopleById(@Path("id") int id);
+    Observable<People> getPeopleById(@Path("id") long id);
 }
